pyproject.toml
[project]
name = "mcp-redfish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "redfish",
    "python-dotenv",
    "fastmcp"
]
